# specs/001-auth-activity-feed/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Media Library API — Auth & Activity Feed
  version: 0.1.0
  description: >
    Contract for feature 001-auth-activity-feed. Covers Google OIDC sign-in,
    session/profile, and the global activity feed (list, post, delete).
    All endpoints are served under the single origin at /api.
servers:
  - url: /api
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dml_session
      description: Signed, httpOnly, Secure session cookie issued on OIDC callback.
...
paths:
...
  /auth/google/callback:
    get:
      summary: OIDC redirect callback
      description: >
        Handles Google's redirect. Validates state/PKCE, upserts the profile by
        google_sub (creating one on first sign-in), creates a session, sets the
        session cookie, and redirects to the SPA. Declined/failed consent
        redirects to the app unauthenticated with an error indicator.
      security: []
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: error, in: query, schema: { type: string } }
      responses:
        '302':
          description: Redirect to the SPA (authenticated on success).
  /auth/logout:
    post:
      summary: Sign out
      description: Deletes the server-side session and clears the cookie.
      security: [{ sessionCookie: [] }]
      responses:
        '204': { description: Session ended. }
...